requirements.txt
absl-py==2.3.1
amqp==5.3.1
asgiref==3.11.1
astunparse==1.6.3
async-timeout==5.0.1
beautifulsoup4==4.14.3
billiard==4.2.4
blinker==1.9.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
cryptography==46.0.5
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
deepface==0.0.98
dlib>=19.24.0
insightface>=0.7.3
onnxruntime>=1.16.0
Django==5.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
exceptiongroup==1.3.1
filelock==3.19.1
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
gast==0.7.0
gdown==5.2.1
google-pasta==0.2.0
grpcio==1.78.1
gunicorn==23.0.0
h5py==3.14.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.10.0
kombu==5.6.2
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==0.1.1
namex==0.1.0
numpy>=2.1.0
opencv-python>=4.13.0
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas>=2.2.0
pillow>=11.0.0
prompt_toolkit==3.0.52
protobuf==6.33.5
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.51.0
accelerate>=1.8.1
ultralytics>=8.0.0
huggingface_hub>=0.20.0
Pygments==2.19.2
PyJWT==2.11.0
pymongo==4.6.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
redis==7.0.1
requests==2.32.5
retina-face==0.0.17
rich==14.3.3
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
scipy>=1.10.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tf_keras==2.20.1
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
Werkzeug==3.1.6
wrapt==2.1.1
zipp==3.23.0
